big refactor!
colocate model-specific graphql code with the models

use sequelize virtual fields for all special model properties (tag totals)

put utility functions in a helper module
