+++ src/db/models/grant.ts
@@ -1,4 +1,9 @@
 import * as Sequelize from 'sequelize';
+import * as decamelize from 'decamelize';
+
+import { Db } from './';
+
+import { ledgerListArgs, MAX_LIMIT } from '../../helpers';
 
 import { OrganizationInstance, OrganizationAttributes } from './organization';
 import { NewsInstance, NewsAttributes } from './news';
@@ -53,54 +58,56 @@ export interface LegacyData {
 export type GrantInstance = Sequelize.Instance<GrantAttributes> &
   GrantAttributes;
 
+const grantColumns = {
+  uuid: {
+    type: Sequelize.UUID,
+    allowNull: true,
+    defaultValue: Sequelize.UUIDV4,
+  },
+  from: {
+    type: Sequelize.INTEGER,
+    allowNull: false,
+    references: { model: 'organization', key: 'id' },
+  },
+  to: {
+    type: Sequelize.INTEGER,
+    allowNull: false,
+    references: { model: 'organization', key: 'id' },
+  },
+  dateFrom: {
+    type: Sequelize.DATEONLY,
+    allowNull: true,
+    field: 'date_from',
+  },
+  dateTo: {
+    type: Sequelize.DATEONLY,
+    allowNull: true,
+    field: 'date_to',
+  },
+  amount: { type: Sequelize.BIGINT, allowNull: true },
+  source: { type: Sequelize.TEXT, allowNull: true },
+  description: { type: Sequelize.TEXT, allowNull: true },
+  internalNotes: {
+    type: Sequelize.TEXT,
+    allowNull: true,
+    field: 'internal_notes',
+  },
+  legacyData: {
+    type: Sequelize.JSON,
+    allowNull: true,
+    field: 'legacy_data',
+  },
+  federalAwardId: {
+    type: Sequelize.STRING,
+    allowNull: true,
+    field: 'federal_award_id',
+  },
+};
+
 export default (sequelize: Sequelize.Sequelize) => {
   let Grant = sequelize.define<GrantInstance, GrantAttributes>(
     'Grant',
-    {
-      uuid: {
-        type: Sequelize.UUID,
-        allowNull: true,
-        defaultValue: Sequelize.UUIDV4,
-      },
-      from: {
-        type: Sequelize.INTEGER,
-        allowNull: false,
-        references: { model: sequelize.models.organization, key: 'id' },
-      },
-      to: {
-        type: Sequelize.INTEGER,
-        allowNull: false,
-        references: { model: sequelize.models.organization, key: 'id' },
-      },
-      dateFrom: {
-        type: Sequelize.DATEONLY,
-        allowNull: true,
-        field: 'date_from',
-      },
-      dateTo: {
-        type: Sequelize.DATEONLY,
-        allowNull: true,
-        field: 'date_to',
-      },
-      amount: { type: Sequelize.BIGINT, allowNull: true },
-      source: { type: Sequelize.TEXT, allowNull: true },
-      description: { type: Sequelize.TEXT, allowNull: true },
-      internalNotes: {
-        type: Sequelize.TEXT,
-        allowNull: true,
-        field: 'internal_notes',
-      },
-      legacyData: {
-        type: Sequelize.JSON,
-        allowNull: true,
-        field: 'legacy_data',
-      },
-      federalAwardId: {
-        type: Sequelize.STRING,
-        allowNull: true,
-        field: 'federal_award_id',
-      },
-    },
+    grantColumns,
     {
       createdAt: 'created_at',
       updatedAt: 'updated_at',
@@ -161,3 +168,51 @@ export default (sequelize: Sequelize.Sequelize) => {
 
   return Grant;
 };
+
+// Add a join to a grant query
+interface AddJoin {
+  (opts: any): string;
+}
+
+export const grantResolver = (db: Db, grantAddWhere?: AddJoin) => async (
+  opts,
+  { limit, offset, orderBy, orderByDirection },
+  context,
+  info
+): Promise<GrantInstance[]> => {
+  let where = '';
+  if (grantAddWhere) {
+    where = grantAddWhere(opts);
+  }
+
+  const results = await db.sequelize.query(
+    `SELECT g.*
+FROM "grant" g
+${where}
+ORDER BY "${decamelize(orderBy)}" ${orderByDirection}
+LIMIT :limit
+OFFSET :offset`,
+    {
+      type: db.Sequelize.QueryTypes.SELECT,
+      model: db.Grant,
+      mapToModel: true,
+      replacements: {
+        limit: Math.min(limit, MAX_LIMIT),
+        offset,
+      },
+    }
+  );
+
+  return results;
+};
+
+export const singleGrantResolver = getId => async (
+  opts,
+  args,
+  context,
+  info
+): Promise<GrantInstance> => {
+  return context.getGrantById.load(getId(opts));
+};
+
+export const grantArgs = ledgerListArgs('Grant', Object.keys(grantColumns));
