+++ src/db/models/organization.ts
@@ -1,4 +1,21 @@
 import * as Sequelize from 'sequelize';
+import { escape } from 'sequelize/lib/sql-string';
+import * as decamelize from 'decamelize';
+import * as DataLoader from 'dataloader';
+
+import { Db } from './';
+
+import {
+  GraphQLFieldConfigMap,
+  GraphQLInt,
+  GraphQLList,
+  GraphQLString,
+} from 'graphql';
+
+import * as GraphQLBigInt from 'graphql-bigint';
+
+import { ledgerListArgs, specialCols, MAX_LIMIT } from '../../helpers';
+
 import { GrantInstance, GrantAttributes } from './grant';
 import { NewsInstance, NewsAttributes } from './news';
 import {
@@ -88,84 +105,82 @@ export interface LegacyData {
 export type OrganizationInstance = Sequelize.Instance<OrganizationAttributes> &
   OrganizationAttributes;
 
+const organizationColumns = {
+  uuid: {
+    type: Sequelize.UUID,
+    allowNull: true,
+    defaultValue: Sequelize.UUIDV4,
+  },
+  name: { type: Sequelize.STRING, allowNull: false },
+  ein: { type: Sequelize.STRING, allowNull: true },
+  duns: { type: Sequelize.STRING, allowNull: true },
+  stateCorpId: {
+    type: Sequelize.STRING,
+    allowNull: true,
+    field: 'state_corp_id',
+  },
+  description: { type: Sequelize.TEXT, allowNull: true },
+  address: { type: Sequelize.JSON, allowNull: true },
+  links: { type: Sequelize.JSON, allowNull: true },
+  founded: { type: Sequelize.DATEONLY, allowNull: true },
+  dissolved: { type: Sequelize.DATEONLY, allowNull: true },
+  legacyData: {
+    type: Sequelize.JSON,
+    allowNull: true,
+    field: 'legacy_data',
+  },
+  publicFunder: {
+    type: Sequelize.BOOLEAN,
+    allowNull: true,
+    field: 'public_funder',
+  },
+
+  // From meta table join:
+  countGrantsFrom: {
+    type: new Sequelize.VIRTUAL(Sequelize.INTEGER),
+    field: 'count_grants_from',
+  },
+  countGrantsTo: {
+    type: new Sequelize.VIRTUAL(Sequelize.INTEGER),
+    field: 'count_grants_to',
+  },
+  countDistinctFunders: {
+    type: new Sequelize.VIRTUAL(Sequelize.INTEGER),
+    field: 'count_distinct_funders',
+  },
+  countDistinctRecipients: {
+    type: new Sequelize.VIRTUAL(Sequelize.INTEGER),
+    field: 'count_distinct_recipients',
+  },
+  totalReceived: {
+    type: new Sequelize.VIRTUAL(Sequelize.BIGINT),
+    field: 'total_received',
+  },
+  totalFunded: {
+    type: new Sequelize.VIRTUAL(Sequelize.BIGINT),
+    field: 'total_funded',
+  },
+  grantdatesStart: {
+    type: new Sequelize.VIRTUAL(Sequelize.DATEONLY),
+    field: 'grantdates_start',
+  },
+  grantdatesEnd: {
+    type: new Sequelize.VIRTUAL(Sequelize.DATEONLY),
+    field: 'grantdates_end',
+  },
+};
+
 export default (sequelize: Sequelize.Sequelize) => {
   let Organization = sequelize.define<
     OrganizationInstance,
     OrganizationAttributes
-  >(
-    'Organization',
-    {
-      uuid: {
-        type: Sequelize.UUID,
-        allowNull: true,
-        defaultValue: Sequelize.UUIDV4,
-      },
-      name: { type: Sequelize.STRING, allowNull: false },
-      ein: { type: Sequelize.STRING, allowNull: true },
-      duns: { type: Sequelize.STRING, allowNull: true },
-      stateCorpId: {
-        type: Sequelize.STRING,
-        allowNull: true,
-        field: 'state_corp_id',
-      },
-      description: { type: Sequelize.TEXT, allowNull: true },
-      address: { type: Sequelize.JSON, allowNull: true },
-      links: { type: Sequelize.JSON, allowNull: true },
-      founded: { type: Sequelize.DATEONLY, allowNull: true },
-      dissolved: { type: Sequelize.DATEONLY, allowNull: true },
-      legacyData: {
-        type: Sequelize.JSON,
-        allowNull: true,
-        field: 'legacy_data',
-      },
-      publicFunder: {
-        type: Sequelize.BOOLEAN,
-        allowNull: true,
-        field: 'public_funder',
-      },
-
-      // From meta table join:
-      countGrantsFrom: {
-        type: new Sequelize.VIRTUAL(Sequelize.INTEGER),
-        field: 'count_grants_from',
-      },
-      countGrantsTo: {
-        type: new Sequelize.VIRTUAL(Sequelize.INTEGER),
-        field: 'count_grants_to',
-      },
-      countDistinctFunders: {
-        type: new Sequelize.VIRTUAL(Sequelize.INTEGER),
-        field: 'count_distinct_funders',
-      },
-      countDistinctRecipients: {
-        type: new Sequelize.VIRTUAL(Sequelize.INTEGER),
-        field: 'count_distinct_recipients',
-      },
-      totalReceived: {
-        type: new Sequelize.VIRTUAL(Sequelize.BIGINT),
-        field: 'total_received',
-      },
-      totalFunded: {
-        type: new Sequelize.VIRTUAL(Sequelize.BIGINT),
-        field: 'total_funded',
-      },
-      grantdatesStart: {
-        type: new Sequelize.VIRTUAL(Sequelize.DATEONLY),
-        field: 'grantdates_start',
-      },
-      grantdatesEnd: {
-        type: new Sequelize.VIRTUAL(Sequelize.DATEONLY),
-        field: 'grantdates_end',
-      },
-    },
-    {
-      createdAt: 'created_at',
-      updatedAt: 'updated_at',
-      underscored: true,
-      freezeTableName: true,
-      tableName: 'organization',
-    }
-  );
+  >('Organization', organizationColumns, {
+    createdAt: 'created_at',
+    updatedAt: 'updated_at',
+    underscored: true,
+    freezeTableName: true,
+    tableName: 'organization',
+  });
 
   // Set up relations
   Organization.associate = ({
@@ -245,4 +260,153 @@ export default (sequelize: Sequelize.Sequelize) => {
   };
 
   return Organization;
+};
+
+/**
+ * These fields are defined as "virtual" in Sequelize
+ * and are the product of a JOIN with the organization_meta table
+ */
+export const organizationSpecialFields: GraphQLFieldConfigMap<never, never> = {
+  countGrantsFrom: { type: GraphQLInt },
+  countGrantsTo: { type: GraphQLInt },
+  countDistinctFunders: { type: GraphQLInt },
+  countDistinctRecipients: { type: GraphQLInt },
+  totalReceived: { type: GraphQLBigInt },
+  totalFunded: { type: GraphQLBigInt },
+  grantdatesStart: { type: GraphQLString },
+  grantdatesEnd: { type: GraphQLString },
+};
+
+// Add a join to a organization query
+interface AddJoin {
+  (opts: any): string;
+}
+
+export const organizationResolver = (db: Db, orgAddJoin?: AddJoin) => async (
+  opts,
+  { limit, offset, orderBy, orderByDirection, nameLike },
+  context,
+  info
+): Promise<OrganizationInstance[]> => {
+  let where = '';
+  if (nameLike) {
+    where = `WHERE o.name ILIKE ${escape(nameLike)}`;
+  }
+
+  let addedJoin = '';
+  if (opts && orgAddJoin) {
+    addedJoin = orgAddJoin(opts);
+  }
+
+  const results = await db.sequelize.query(
+    `SELECT o.*, ${specialCols(organizationSpecialFields, 'om').join(',')}
+FROM organization o
+LEFT JOIN organization_meta om ON o.id=om.id
+${addedJoin}
+${where}
+ORDER BY "${decamelize(orderBy)}" ${orderByDirection}
+LIMIT :limit
+OFFSET :offset`,
+    {
+      type: db.Sequelize.QueryTypes.SELECT,
+      model: db.Organization,
+      mapToModel: true,
+      replacements: {
+        limit: Math.min(limit, MAX_LIMIT),
+        offset,
+      },
+    }
+  );
+
+  return results;
+};
+
+export const createGetOrganizationByIdDataloader = (
+  db: Db
+): DataLoader<number, OrganizationInstance> => {
+  return new DataLoader(
+    async (ids: number[]): Promise<OrganizationInstance[]> => {
+      const results = await db.sequelize.query(
+        `SELECT o.*, ${specialCols(organizationSpecialFields, 'om').join(',')}
+  FROM organization o
+  LEFT JOIN organization_meta om ON o.id=om.id
+  WHERE o.id IN(:ids)`,
+        {
+          type: db.Sequelize.QueryTypes.SELECT,
+          model: db.Organization,
+          mapToModel: true,
+          replacements: { ids },
+        }
+      );
+
+      const ordered = ids.map(
+        id =>
+          results.find(o => o.id === id) ||
+          new Error(`cannot find organization with id ${id}`)
+      );
+
+      return ordered;
+    }
+  );
+};
+
+export const createGetOrganizationByUuidDataloader = (
+  db: Db
+): DataLoader<number, OrganizationInstance> => {
+  return new DataLoader(
+    async (uuids: number[]): Promise<OrganizationInstance[]> => {
+      const results = await db.sequelize.query(
+        `SELECT o.*, ${specialCols(organizationSpecialFields, 'om').join(',')}
+  FROM organization o
+  LEFT JOIN organization_meta om ON o.id=om.id
+  WHERE o.uuid IN(:uuids)`,
+        {
+          type: db.Sequelize.QueryTypes.SELECT,
+          model: db.Organization,
+          mapToModel: true,
+          replacements: { uuids },
+        }
+      );
+
+      const ordered = uuids.map(
+        uuid =>
+          results.find(o => o.uuid === uuid) ||
+          new Error(`cannot find organization with uuid ${uuid}`)
+      );
+
+      return ordered;
+    }
+  );
+};
+
+interface GetIdFromOpts {
+  (opts: any): number;
+}
+
+export const singleOrganizationResolver = (getId?: GetIdFromOpts) => async (
+  opts,
+  args,
+  context,
+  info
+): Promise<OrganizationInstance> => {
+  if (getId) {
+    return context.getOrganizationById.load(getId(opts));
+  }
+
+  if (args.id) {
+    return context.getOrganizationById.load(args.id);
+  }
+
+  if (args.uuid) {
+    return context.getOrganizationByUuid.load(args.uuid);
+  }
+
+  throw new Error('must supply either getId or args with id or uuid property');
+};
+
+export const organizationArgs = {
+  ...ledgerListArgs('Organization', Object.keys(organizationColumns)),
+  nameLike: {
+    type: GraphQLString,
+  },
 };
