+++ src/server.ts
@@ -1,8 +1,6 @@
 import * as Sequelize from 'sequelize';
-import * as decamelize from 'decamelize';
 import { GraphQLServer } from 'graphql-yoga';
 import { createContext, EXPECTED_OPTIONS_KEY } from 'dataloader-sequelize';
-import * as DataLoader from 'dataloader';
 import {
   resolver,
   attributeFields,
@@ -14,10 +12,8 @@ import {
   graphql,
   GraphQLSchema,
   GraphQLObjectType,
-  GraphQLString,
   GraphQLInt,
   GraphQLList,
-  GraphQLEnumType,
 } from 'graphql';
 
 import { escape } from 'sequelize/lib/sql-string';
@@ -26,42 +22,38 @@ import * as GraphQLBigInt from 'graphql-bigint';
 
 import { Db } from './db/models';
 
-import { OrganizationInstance } from './db/models/organization';
-
-const MAX_LIMIT = 100;
-
-const organizationSpecialFields = {
-  countGrantsFrom: { type: GraphQLInt },
-  countGrantsTo: { type: GraphQLInt },
-  countDistinctFunders: { type: GraphQLInt },
-  countDistinctRecipients: { type: GraphQLInt },
-  totalReceived: { type: GraphQLBigInt },
-  totalFunded: { type: GraphQLBigInt },
-  grantdatesStart: { type: GraphQLString },
-  grantdatesEnd: { type: GraphQLString },
-};
+import {
+  organizationResolver,
+  organizationArgs,
+  organizationSpecialFields,
+  singleOrganizationResolver,
+  createGetOrganizationByIdDataloader,
+  createGetOrganizationByUuidDataloader,
+} from './db/models/organization';
+import { grantResolver, grantArgs } from './db/models/grant';
+import {
+  grantTagResolver,
+  grantTagArgs,
+  grantTagSpecialFields,
+} from './db/models/grantTag';
+import {
+  organizationTagResolver,
+  organizationTagArgs,
+  organizationTagSpecialFields,
+} from './db/models/organizationTag';
+import {
+  nteeGrantTypeResolver,
+  nteeGrantTypeArgs,
+  nteeGrantTypeSpecialFields,
+} from './db/models/nteeGrantType';
+import {
+  nteeOrganizationTypeResolver,
+  nteeOrganizationTypeArgs,
+  nteeOrganizationTypeSpecialFields,
+} from './db/models/nteeOrganizationType';
 
 export default function createServer(db: Db): GraphQLServer {
   resolver.contextToOptions = { [EXPECTED_OPTIONS_KEY]: EXPECTED_OPTIONS_KEY };
-
-  // Arguments
-  const grantTagArgs = ledgerListArgs(db.GrantTag, ['total']);
-  const nteeGrantTypeArgs = ledgerListArgs(db.NteeGrantType, ['total']);
-  const nteeOrganizationTypeArgs = ledgerListArgs(db.NteeOrganizationType, [
-    'totalFunded',
-    'totalReceived',
-  ]);
-  const organizationTagArgs = ledgerListArgs(db.OrganizationTag, [
-    'totalFunded',
-    'totalReceived',
-  ]);
-  const grantArgs = ledgerListArgs(db.Grant);
-  const organizationArgs = {
-    ...ledgerListArgs(db.Organization, Object.keys(organizationSpecialFields)),
-    nameLike: {
-      type: GraphQLString,
-    },
-  };
 
   // Types
   const shallowOrganizationType = new GraphQLObjectType({
@@ -75,8 +67,7 @@ export default function createServer(db: Db): GraphQLServer {
     description: 'Tag associated with an organization',
     fields: {
       ...attributeFields(db.OrganizationTag, { exclude: ['id'] }),
-      totalFunded: { type: GraphQLBigInt },
-      totalReceived: { type: GraphQLBigInt },
+      ...organizationTagSpecialFields,
     },
   });
 
@@ -85,7 +76,7 @@ export default function createServer(db: Db): GraphQLServer {
     description: 'Tag associated with a grant',
     fields: {
       ...attributeFields(db.GrantTag, { exclude: ['id'] }),
-      total: { type: GraphQLBigInt },
+      ...grantTagSpecialFields,
     },
   });
 
@@ -94,7 +85,7 @@ export default function createServer(db: Db): GraphQLServer {
     description: 'NTEE classification of a grant',
     fields: {
       ...attributeFields(db.NteeGrantType, { exclude: ['id'] }),
-      total: { type: GraphQLBigInt },
+      ...nteeGrantTypeSpecialFields,
     },
   });
 
@@ -103,8 +94,7 @@ export default function createServer(db: Db): GraphQLServer {
     description: 'NTEE classification of an organization',
     fields: {
       ...attributeFields(db.NteeOrganizationType, { exclude: ['id'] }),
-      totalFunded: { type: GraphQLBigInt },
-      totalReceived: { type: GraphQLBigInt },
+      ...nteeOrganizationTypeSpecialFields,
     },
   });
 
@@ -345,391 +335,3 @@ export default function createServer(db: Db): GraphQLServer {
     },
   });
 }
-
-interface OrganizationTagResolverOptions {
-  limitToOrganizationId: boolean;
-}
-
-const defaultOrganizationTagResolverOptions = {
-  limitToOrganizationId: false,
-};
-
-interface GrantTagResolverOptions {
-  limitToGrantId: boolean;
-}
-
-const defaultGrantTagResolverOptions = {
-  limitToGrantId: false,
-};
-
-// Add a join to a organization query
-interface AddJoin {
-  (opts: any): string;
-}
-
-const grantResolver = (db, grantAddWhere?: AddJoin) => async (
-  opts,
-  { limit, offset, orderBy, orderByDirection },
-  context,
-  info
-) => {
-  let where = '';
-  if (grantAddWhere) {
-    where = grantAddWhere(opts);
-  }
-
-  const results = await db.sequelize.query(
-    `SELECT g.*
-FROM "grant" g
-${where}
-ORDER BY "${decamelize(orderBy)}" ${orderByDirection}
-LIMIT :limit
-OFFSET :offset`,
-    {
-      type: db.Sequelize.QueryTypes.SELECT,
-      model: db.Grant,
-      mapToModel: true,
-      replacements: {
-        limit: Math.min(limit, MAX_LIMIT),
-        offset,
-      },
-    }
-  );
-
-  return results;
-};
-
-const metaCols = Object.keys(organizationSpecialFields).map(
-  col => `om.${decamelize(col)} AS "${decamelize(col)}"`
-);
-
-const createGetOrganizationByIdDataloader = (db: Db) => {
-  return new DataLoader(
-    async (ids: number[]): Promise<OrganizationInstance[]> => {
-      const results = await db.sequelize.query(
-        `SELECT o.*, ${metaCols.join(',')}
-  FROM organization o
-  LEFT JOIN organization_meta om ON o.id=om.id
-  WHERE o.id IN(:ids)`,
-        {
-          type: db.Sequelize.QueryTypes.SELECT,
-          model: db.Organization,
-          mapToModel: true,
-          replacements: { ids },
-        }
-      );
-
-      const ordered = ids.map(
-        id =>
-          results.find(o => o.id === id) ||
-          new Error(`cannot find organization with id ${id}`)
-      );
-
-      return ordered;
-    }
-  );
-};
-
-const createGetOrganizationByUuidDataloader = (db: Db) => {
-  return new DataLoader(
-    async (uuids: number[]): Promise<OrganizationInstance[]> => {
-      const results = await db.sequelize.query(
-        `SELECT o.*, ${metaCols.join(',')}
-  FROM organization o
-  LEFT JOIN organization_meta om ON o.id=om.id
-  WHERE o.uuid IN(:uuids)`,
-        {
-          type: db.Sequelize.QueryTypes.SELECT,
-          model: db.Organization,
-          mapToModel: true,
-          replacements: { uuids },
-        }
-      );
-
-      const ordered = uuids.map(
-        uuid =>
-          results.find(o => o.uuid === uuid) ||
-          new Error(`cannot find organization with uuid ${uuid}`)
-      );
-
-      return ordered;
-    }
-  );
-};
-
-interface GetIdFromOpts {
-  (opts: any): number;
-}
-
-const singleOrganizationResolver = (getId?: GetIdFromOpts) => async (
-  opts,
-  args,
-  context,
-  info
-) => {
-  if (getId) {
-    return context.getOrganizationById.load(getId(opts));
-  }
-
-  if (args.id) {
-    return context.getOrganizationById.load(args.id);
-  }
-
-  if (args.uuid) {
-    return context.getOrganizationByUuid.load(args.uuid);
-  }
-};
-
-const singleGrantResolver = getId => async (opts, args, context, info) => {
-  return context.getGrantById.load(getId(opts));
-};
-
-const organizationResolver = (db, orgAddJoin?: AddJoin) => async (
-  opts,
-  { limit, offset, orderBy, orderByDirection, nameLike },
-  context,
-  info
-) => {
-  let where = '';
-  if (nameLike) {
-    where = `WHERE o.name ILIKE ${escape(nameLike)}`;
-  }
-
-  let addedJoin = '';
-  if (opts && orgAddJoin) {
-    addedJoin = orgAddJoin(opts);
-  }
-
-  const results = await db.sequelize.query(
-    `SELECT o.*, ${metaCols.join(',')}
-FROM organization o
-LEFT JOIN organization_meta om ON o.id=om.id
-${addedJoin}
-${where}
-ORDER BY "${decamelize(orderBy)}" ${orderByDirection}
-LIMIT :limit
-OFFSET :offset`,
-    {
-      type: db.Sequelize.QueryTypes.SELECT,
-      model: db.Organization,
-      mapToModel: true,
-      replacements: {
-        limit: Math.min(limit, MAX_LIMIT),
-        offset,
-      },
-    }
-  );
-
-  return results;
-};
-
-const organizationTagResolver = (
-  db,
-  resolverOpts: OrganizationTagResolverOptions = defaultOrganizationTagResolverOptions
-) => async (
-  opts,
-  { limit, offset, orderBy, orderByDirection, uuid = null },
-  context,
-  info
-) => {
-  let where = '';
-  // Fetching only organization tags related to a specific organization
-  if (resolverOpts.limitToOrganizationId) {
-    where = `WHERE oot.organization_id=${escape(opts.dataValues.id)}`;
-  } else {
-    where = uuid ? `WHERE ot.uuid = ${escape(uuid)}` : '';
-  }
-
-  const results = await db.sequelize.query(
-    `SELECT ot.id, ot.uuid, ot.name, ot.description, SUM(gf.amount) as "totalFunded", SUM(gr.amount) as "totalReceived"
-FROM organization_tag ot
-LEFT JOIN organization_organization_tag oot ON ot.id=oot.organization_tag_id
-LEFT JOIN "grant" gf ON gf.from=oot.organization_id
-LEFT JOIN "grant" gr ON gr.to=oot.organization_id
-${where}
-GROUP BY ot.id
-ORDER BY "${orderBy}" ${orderByDirection}
-LIMIT :limit
-OFFSET :offset`,
-    {
-      type: db.Sequelize.QueryTypes.SELECT,
-      replacements: {
-        limit: Math.min(limit, MAX_LIMIT),
-        offset,
-      },
-    }
-  );
-
-  return results;
-};
-
-const nteeOrganizationTypeResolver = (
-  db,
-  resolverOpts: OrganizationTagResolverOptions = defaultOrganizationTagResolverOptions
-) => async (
-  opts,
-  { limit, offset, orderBy, orderByDirection, uuid = null },
-  context,
-  info
-) => {
-  let where = '';
-  // Fetching only organization tags related to a specific organization
-  if (resolverOpts.limitToOrganizationId) {
-    where = `WHERE oot.organization_id=${escape(opts.dataValues.id)}`;
-  } else {
-    where = uuid ? `WHERE ot.uuid = ${escape(uuid)}` : '';
-  }
-
-  const results = await db.sequelize.query(
-    `SELECT ot.id, ot.uuid, ot.name, ot.code, ot.description, SUM(gf.amount) as "totalFunded", SUM(gr.amount) as "totalReceived"
-FROM ntee_organization_type ot
-LEFT JOIN organization_ntee_organization_type oot ON ot.id=oot.ntee_organization_type_id
-LEFT JOIN "grant" gf ON gf.from=oot.organization_id
-LEFT JOIN "grant" gr ON gr.to=oot.organization_id
-${where}
-GROUP BY ot.id
-ORDER BY "${orderBy}" ${orderByDirection}
-LIMIT :limit
-OFFSET :offset`,
-    {
-      type: db.Sequelize.QueryTypes.SELECT,
-      replacements: {
-        limit: Math.min(limit, MAX_LIMIT),
-        offset,
-      },
-    }
-  );
-
-  return results;
-};
-
-const grantTagResolver = (
-  db,
-  resolverOpts: GrantTagResolverOptions = defaultGrantTagResolverOptions
-) => async (
-  opts,
-  { limit, offset, orderBy, orderByDirection, uuid = null },
-  context,
-  info
-) => {
-  let where = '';
-  // Fetching only grant tags related to a specific grant
-  if (resolverOpts.limitToGrantId) {
-    where = `WHERE g.id=${escape(opts.dataValues.id)}`;
-  } else {
-    where = uuid ? `WHERE gt.uuid = ${escape(uuid)}` : '';
-  }
-
-  const results = await db.sequelize.query(
-    `SELECT gt.id, gt.uuid, gt.name, gt.description, SUM(g.amount) as total
-FROM grant_tag gt
-LEFT JOIN grant_grant_tag ggt ON gt.id=ggt.grant_tag_id
-LEFT JOIN "grant" g ON ggt.grant_id=g.id
-${where}
-GROUP BY gt.id
-ORDER BY "${orderBy}" ${orderByDirection}
-LIMIT :limit
-OFFSET :offset`,
-    {
-      type: db.Sequelize.QueryTypes.SELECT,
-      replacements: {
-        limit: Math.min(limit, MAX_LIMIT),
-        offset,
-      },
-    }
-  );
-
-  return results;
-};
-
-const nteeGrantTypeResolver = (
-  db,
-  resolverOpts: GrantTagResolverOptions = defaultGrantTagResolverOptions
-) => async (
-  opts,
-  { limit, offset, orderBy, orderByDirection, uuid = null },
-  context,
-  info
-) => {
-  let where = '';
-  // Fetching only grant tags related to a specific grant
-  if (resolverOpts.limitToGrantId) {
-    where = `WHERE g.id=${escape(opts.dataValues.id)}`;
-  } else {
-    where = uuid ? `WHERE gt.uuid = ${escape(uuid)}` : '';
-  }
-
-  const results = await db.sequelize.query(
-    `SELECT gt.id, gt.uuid, gt.name, gt.code, gt.description, SUM(g.amount) as total
-FROM ntee_grant_type gt
-LEFT JOIN grant_ntee_grant_type ggt ON gt.id=ggt.ntee_grant_type_id
-LEFT JOIN "grant" g ON ggt.grant_id=g.id
-${where}
-GROUP BY gt.id
-ORDER BY "${orderBy}" ${orderByDirection}
-LIMIT :limit
-OFFSET :offset`,
-    {
-      type: db.Sequelize.QueryTypes.SELECT,
-      replacements: {
-        limit: Math.min(limit, MAX_LIMIT),
-        offset,
-      },
-    }
-  );
-
-  return results;
-};
-
-const ledgerListArgs = (
-  model: Sequelize.Model<any, any>,
-  orderBySpecialCols: string[] = []
-) => ({
-  orderBy: {
-    type: new GraphQLEnumType({
-      name: `orderBy${model.name}`,
-      // @ts-ignore tableAttributes is not in sequelize type defs
-      values: Object.keys(model.tableAttributes).reduce(
-        (acc, cur) => ({
-          ...acc,
-          [cur]: { value: cur },
-        }),
-        orderBySpecialCols.reduce(
-          (acc, cur) => ({
-            ...acc,
-            [cur]: { value: cur },
-          }),
-          {}
-        )
-      ),
-    }),
-    defaultValue: 'id',
-    description: 'sort results by given field',
-  },
-  orderByDirection: {
-    type: new GraphQLEnumType({
-      name: `orderByDirection${model.name}`,
-      values: {
-        ASC: { value: 'ASC NULLS LAST' },
-        DESC: { value: 'DESC NULLS LAST' },
-      },
-    }),
-    defaultValue: 'ASC NULLS LAST',
-    description: 'sort direction',
-  },
-  limit: {
-    type: GraphQLInt,
-    defaultValue: 10,
-    description: `Number of items to return, maximum ${MAX_LIMIT}`,
-  },
-  offset: {
-    type: GraphQLInt,
-    defaultValue: 0,
-  },
-  uuid: {
-    type: GraphQLString,
-  },
-  id: {
-    type: GraphQLInt,
-  },
-});
